pyproject.toml
[tool.ruff]
# Allow lines to be as long as 120 characters.
line-length = 120

[tool.poetry]
name = "django-api"
version = "0.0.1"
description = "Django API template"
authors = ["Sergey Zahranychnyi <zagranlab@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "~=5.0.1"
djangorestframework = "^3.14.0"
asgiref = "^3.7.2"
gunicorn = "^21.2.0"
packaging = "^23.2"
python-dotenv = "^1.0.0"
sqlparse = "^0.4.4"
typing-extensions = "^4.9.0"
whitenoise = "^6.6.0"
virtualenv = "^20.25.0"
dumb-init = "^1.2.5.post1"
drf-spectacular = "^0.27.1"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-django = "^4.8.0"
pytest-factoryboy = "^2.6.0"
pytest-mock = "^3.12.0"
factory-boy = "^3.3.0"
setuptools = "^69.0.3"

[tool.ruff.mccabe]
max-complexity = 18


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
